Move some device management modules into hardware_set/
